dev: custom hook to fetch comments
